Moved CustomFonts to Classes folder/namespace
